+++ tests/utils/plugin_signer.py
@@ -82,15 +82,3 @@ class PluginSigner:
 
         selector = get_selector_from_name(selector_name)
         return await self.account.readOnPlugin(plugin, selector, arguments).call()
-
-
-class StarkPluginSigner(PluginSigner):
-    def __init__(self, stark_key: StarkKeyPair, account: StarknetContract, plugin_address):
-        super().__init__(account, plugin_address)
-        self.stark_key = stark_key
-        self.public_key = stark_key.public_key
-
-    def sign(self, message_hash: int) -> List[int]:
-        return [self.plugin_address] + list(sign(msg_hash=message_hash, priv_key=self.stark_key.private_key))
-
-
